+++ mobile/android/android-components/components/browser/toolbar/src/test/java/mozilla/components/browser/toolbar/edit/EditToolbarTest.kt
@@ -4,6 +4,7 @@
 
 package mozilla.components.browser.toolbar.edit
 
+import kotlinx.coroutines.runBlocking
 import mozilla.components.browser.toolbar.BrowserToolbar
 import mozilla.components.concept.toolbar.AutocompleteDelegate
 import org.junit.Assert.assertEquals
@@ -13,20 +14,29 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.RuntimeEnvironment
+import java.util.concurrent.CountDownLatch
 
 @RunWith(RobolectricTestRunner::class)
 class EditToolbarTest {
     @Test
-    fun `entered text is forwarded to autocomplete filter`() {
+    fun `entered text is forwarded to async autocomplete filter`() {
         val toolbar = BrowserToolbar(RuntimeEnvironment.application)
         toolbar.editToolbar.urlView.onAttachedToWindow()
 
+        val latch = CountDownLatch(1)
         var invokedWithParams: List<Any?>? = null
         toolbar.setAutocompleteListener { p1, p2 ->
             invokedWithParams = listOf(p1, p2)
+            latch.countDown()
         }
 
         toolbar.editToolbar.urlView.setText("Hello")
+
+        // Autocomplete filter will be invoked on a worker thread.
+        // Serialize here for the sake of tests.
+        runBlocking {
+            latch.await()
+        }
 
         assertEquals("Hello", invokedWithParams!![0])
         assertTrue(invokedWithParams!![1] is AutocompleteDelegate)